+++ packages/cli/doc-core/src/shared/types/defaultTheme.ts
@@ -115,6 +115,7 @@ export type NavItemWithLink = {
   text: string;
   link: string;
   activeMatch?: string;
+  position?: 'left' | 'right';
 };
 
 export type NavItemChildren = {
@@ -125,6 +126,7 @@ export type NavItemChildren = {
 export interface NavItemWithChildren {
   text?: string;
   items: (NavItemWithChildren | NavItemWithLink)[];
+  position?: 'left' | 'right';
 }
 
 // image -----------------------------------------------------------------------