+++ tests/integration/builder-rspack-with-ssr/src/App.tsx
@@ -1,9 +1,0 @@
-import './App.scss';
-
-const App = () => (
-  <p className="description">
-    Get started by editing <code className="code">src/App.tsx</code>
-  </p>
-);
-
-export default App;