+++ tests/integration/server-hook/template/test/index.test.ts
@@ -1,20 +1,25 @@
 import path from 'path';
 import { Page } from 'puppeteer';
-import { launchApp, getPort, killApp } from '../../../../utils/modernTestUtils';
+import {
+  launchApp,
+  getPort,
+  killApp,
+  openPage,
+} from '../../../../utils/modernTestUtils';
 
 const appPath = path.resolve(__dirname, '../');
-
-declare let page: Page;
 
 describe('test status code page', () => {
   let app: any;
   let port: number;
+  let page: Page;
 
   beforeAll(async () => {
     jest.setTimeout(1000 * 60 * 2);
     port = await getPort();
 
     app = await launchApp(appPath, port);
+    page = await openPage();
   });
 
   afterAll(async () => {
@@ -25,7 +30,7 @@ describe('test status code page', () => {
 
   it('should template api work correctly ', async () => {
     const response = await page.goto(`http://localhost:${port}`);
-    const text = await response.text();
+    const text = await response!.text();
 
     expect(text).toMatch('<meta name="text-append" content="hello modern">');
     expect(text).toMatch('<meta name="text-prepend" content="hello modern">');