+++ packages/cli/babel-preset-base/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/babel-preset-base
 
+## 2.17.1
+
+### Patch Changes
+
+- @modern-js/utils@2.17.1
+
 ## 2.17.0
 
 ### Patch Changes