+++ packages/document/doc-tools-doc/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/doc-tools-doc
 
+## 2.17.1
+
+### Patch Changes
+
+- 846590d: feat: add new modern.js doc support edit this page on github (#3527)
+
+  feat: 添加新的 modern.js 文档支持在 Github 上编辑此页面 (#3527)。
+
+- e3733a3: feat(doc-core): doc plugin system doc optimization
+
+  feat(doc-core): 文档框架插件机制文档优化
+
 ## 2.17.0
 
 ## 2.16.0