+++ packages/generator/generators/test-generator/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @modern-js/test-generator
 
+## 3.1.12
+
 ## 3.1.11
 
 ## 3.1.10