+++ packages/solutions/monorepo-tools/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @modern-js/monorepo-tools
 
+## 2.17.1
+
+### Patch Changes
+
+- @modern-js/core@2.17.1
+- @modern-js/plugin-changeset@2.17.1
+- @modern-js/plugin-i18n@2.17.1
+- @modern-js/plugin-lint@2.17.1
+- @modern-js/new-action@2.17.1
+- @modern-js/plugin@2.17.1
+- @modern-js/upgrade@2.17.1
+- @modern-js/utils@2.17.1
+
 ## 2.17.0
 
 ### Patch Changes